pyproject.toml
[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
packages = [
  { include = "knowledge_flow_app" }
]

[tool.poetry]
name = "knowledge_flow_app"
version = "0.1-dev"
description = "A knowledge flow backend to manage and serve knowledge base."
authors = ["Dorian Finel <Dorian.finel-bacha.e@thalesdigital.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
# Framework & Server
fastapi = "^0.115.7"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
azure-identity = "^1.19.0"
# OpenAI / IA
openai = "^1.60.0"
# LangChain / LangGraph (si nécessaire)
langchain = "^0.3.25"      
langchain-core = "^0.3.31"
langchain-community = "^0.3.15"
langchain-openai = "^0.3.1"
coloredlogs = "^15.0.1"    
# YAML pour parse_server_configuration
PyYAML = "^6.0"
# Pydantic (FastAPI  + structuration)
pydantic = ">=2.5.2,<3.0.0"
pydantic-settings = "^2.7.1"
##
ipython = "^8.31.0"
python-multipart = "^0.0.20"
opensearch-py = "^2.8.0"
python-docx = "^1.1.2"
minio = "^7.2.15"
pypandoc-binary = "^1.15"
pypandoc = "^1.15"
pandas = "^2.2.3"
pypdf = "^5.4.0"
python-pptx = "^1.0.2"
google-cloud-storage = "^3.1.0"
rich = "^14.0.0"
shortuuid = "^1.0.13"
openpyxl = "^3.1.5"
docling = "^2.31.2"
hf-xet = "^1.1.1"
torch = "2.2.2"

[tool.poetry.scripts]
chatbot = "knowledge_flow_app.main:main"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
reportlab = "^4.3.1"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^23.1.0"
isort = "^5.10.1"

